Add Evaluate boolean expressions.
The pcode is more less generated correctly. But there's still something important missing: I had to make a distiction between directions and constans when evaluating expressions... and whenever there's an =, check first if that E is constant or the result of an evaluation. That way I'll know if the assignment is var-to-var or const-to-var

Signed-off-by: Marcela Rosales <marcelarosalesj@gmail.com>