Added Dimensional's variables support
No need of Dimensions Table. Instead I added a node pointers to the Elements' table. If it is pointing to  NULL, then the variable is simple, if it is pointing to a node, then it has at least one dimension. The node has information about the dimension and it's position in memory.

Signed-off-by: Marcela Rosales <marcelarosalesj@gmail.com>